requirements.txt
torch>=2.2,<3.0
torchvision>=0.17,<1.0
hydra-core>=1.3,<2.0
tensorboard>=2.16,<3.0
pandas>=2.0,<3.0
numpy>=1.26,<3.0
tqdm>=4.66,<5.0
